Classes refactored and moved to it own files
