+++ src/APIM_ARMTemplate/apimtemplate/apimtemplate.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp2.1</TargetFramework>
-    <ProjectGuid>C575A23E-CCED-4F0F-B3AC-BE8E450CE3F0</ProjectGuid>
+    <ProjectGuid>{B5183465-2BC1-4206-9C9F-5AC9615AC941}</ProjectGuid>
   </PropertyGroup>
 
   <ItemGroup>
